+++ pkg/cms/funcmap.go
@@ -4,6 +4,8 @@ import (
 	"bytes"
 	"fmt"
 	"github.com/Masterminds/sprig/v3"
+	"github.com/emvi/shifu/pkg/admin/db"
+	"github.com/emvi/shifu/pkg/admin/model"
 	"github.com/emvi/shifu/pkg/cfg"
 	"github.com/yuin/goldmark"
 	"github.com/yuin/goldmark/extension"
@@ -11,7 +13,6 @@ import (
 	"html/template"
 	"log/slog"
 	"math/rand/v2"
-	"net/http"
 	"os"
 	"path/filepath"
 	"strconv"
@@ -267,18 +268,28 @@ func formatInt(i int) string {
 	return out[:len(out)-1]
 }
 
-func isLoggedIn(r *http.Request) bool {
-	cookie, err := r.Cookie("session")
+func isLoggedIn(page *Content) bool {
+	cookie, err := page.Request.Cookie("session")
 
-	if err != nil {
+	if err != nil || cookie.Value == "" {
 		return false
 	}
 
-	return cookie.Value != "" && cookie.Expires.Before(time.Now())
+	s := new(model.Session)
+
+	if err := db.Get().Get(s, `SELECT * FROM "session" WHERE session = ?`, cookie.Value); err != nil {
+		return false
+	}
+
+	if s == nil || s.Expires.Before(time.Now()) {
+		return false
+	}
+
+	return true
 }
 
-func adminHead(r *http.Request) template.HTML {
-	if isLoggedIn(r) {
+func adminHead(page *Content) template.HTML {
+	if isLoggedIn(page) {
 		path := cfg.Get().UI.Path
 		return template.HTML(fmt.Sprintf(`<link rel="prefetch" href="%s/static/fonts/Inter-Medium.woff2" as="font" type="font/woff2" />
 			<link rel="prefetch" href="%s/static/fonts/Inter-Regular.woff2" as="font" type="font/woff2" />
@@ -297,7 +308,7 @@ func adminHead(r *http.Request) template.HTML {
 }
 
 func adminBody(page *Content) template.HTML {
-	if isLoggedIn(page.Request) {
+	if isLoggedIn(page) {
 		path := cfg.Get().UI.Path
 		return template.HTML(fmt.Sprintf(`<div hx-get="%s/toolbar?path=%s&language=%s"
 			hx-swap="outerHTML"