+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 1.2.0
+
+* added `Content.Clone()` method
+* added file name sanitization
+* changed `adminHead` and `loggedIn` parameters to use the page instead of the request directly
+* fixed checking if user is logged in for the admin UI
+* fixed data select fields when the key is different from the value
+* fixed tree state not being applied to individual pages
+* fixed editing media file names
+* fixed URL encoding paths for file synchronization
+* updated dependencies
+
 ## 1.1.2
 
 * improved synchronization mutexes